WIP: add tests to open lineage panel
